test: replace hash scroll source snapshot
